EES-6465 Separate Publications search into new controller and service
